리팩토링 : Winner 클래스 분리
