+++ src/main/java/domain/RacingGame.java
@@ -12,7 +12,7 @@ public class RacingGame {
 	private List<Car> carList;
 	
 	/* 우승자 이름을 저장 */
-	private String winner = "";
+	private Winner winner;
 	
 	/* 유저로 부터 입력을 받기위한 InputManager 인스턴스 */
 	private InputManager inputManager;
@@ -20,6 +20,7 @@ public class RacingGame {
 	public RacingGame() {
 		carList = new ArrayList<>();
 		inputManager = new InputManager();
+		winner = new Winner();
 	}
 
 	/*
@@ -29,7 +30,7 @@ public class RacingGame {
 	public void play() {
 		prepareRacing();
 		raceCars();
-		showWinner();
+		winner.showWinner(carList);
 	}
 
 	/*
@@ -55,32 +56,6 @@ public class RacingGame {
 			}
 			System.out.println();
 		}
-	}
-	
-	/*
-	 * 우승자를 계산해서 출력하는 메서드
-	 * 가장 높은 position에 해당하는 carName을 출력한다.
-	 */
-	private void showWinner() {
-		int MAX = 0;
-		for (Car car : carList) {
-			MAX = updateMAX(MAX, car);
-			if (MAX == car.getPosition() && winner != car.getName()) {
-				winner += ", " + car.getName();
-			}
-		}
-		System.out.println(winner + "가 최종 우승했습니다.");
-	}
-
-	/*
-	 * MAX값이 업데이트 되면 winner변수도 같이 업데이트시키는 메서드
-	 */
-	private int updateMAX(int MAX, Car car) {
-		if (MAX < car.getPosition()) {
-			winner = car.getName();
-			return car.getPosition();
-		}
-		return MAX;
 	}
 	
 }
